edit: move extension to misc/vscodeextensions
