+++ extractor/filesystem/list/list.go
@@ -60,6 +60,7 @@ import (
 	"github.com/google/osv-scalibr/extractor/filesystem/language/swift/packageresolved"
 	"github.com/google/osv-scalibr/extractor/filesystem/language/swift/podfilelock"
 	"github.com/google/osv-scalibr/extractor/filesystem/language/wordpress/plugins"
+	"github.com/google/osv-scalibr/extractor/filesystem/misc/vscodeextensions"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/apk"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/cos"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/dpkg"
@@ -73,7 +74,6 @@ import (
 	"github.com/google/osv-scalibr/extractor/filesystem/os/portage"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/rpm"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/snap"
-	vscodeextensions "github.com/google/osv-scalibr/extractor/filesystem/os/vscode/extensions"
 	"github.com/google/osv-scalibr/extractor/filesystem/sbom/cdx"
 	"github.com/google/osv-scalibr/extractor/filesystem/sbom/spdx"
 	"github.com/google/osv-scalibr/plugin"
@@ -177,19 +177,22 @@ var (
 
 	// OS extractors.
 	OS = InitMap{
-		dpkg.Name:             {dpkg.NewDefault},
-		apk.Name:              {apk.NewDefault},
-		rpm.Name:              {rpm.NewDefault},
-		cos.Name:              {cos.NewDefault},
-		snap.Name:             {snap.NewDefault},
-		nix.Name:              {nix.New},
-		module.Name:           {module.NewDefault},
-		vmlinuz.Name:          {vmlinuz.NewDefault},
-		pacman.Name:           {pacman.NewDefault},
-		portage.Name:          {portage.NewDefault},
-		flatpak.Name:          {flatpak.NewDefault},
-		homebrew.Name:         {homebrew.New},
-		macapps.Name:          {macapps.NewDefault},
+		dpkg.Name:     {dpkg.NewDefault},
+		apk.Name:      {apk.NewDefault},
+		rpm.Name:      {rpm.NewDefault},
+		cos.Name:      {cos.NewDefault},
+		snap.Name:     {snap.NewDefault},
+		nix.Name:      {nix.New},
+		module.Name:   {module.NewDefault},
+		vmlinuz.Name:  {vmlinuz.NewDefault},
+		pacman.Name:   {pacman.NewDefault},
+		portage.Name:  {portage.NewDefault},
+		flatpak.Name:  {flatpak.NewDefault},
+		homebrew.Name: {homebrew.New},
+		macapps.Name:  {macapps.NewDefault},
+	}
+
+	Misc = InitMap{
 		vscodeextensions.Name: {vscodeextensions.New},
 	}
 
@@ -216,6 +219,7 @@ var (
 		SBOM,
 		Swift,
 		OS,
+		Misc,
 		Containers,
 		Wordpress,
 	)
