+++ extractor/filesystem/misc/vscodeextensions/testdata/extensions-windows.json
