+++ extractor/filesystem/misc/vscodeextensions/testdata/extensions.json
