+++ extractor/filesystem/misc/vscodeextensions/testdata/invalid.json
