+++ extractor/filesystem/misc/vscodeextensions/testdata/one-extension.json
